+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,8 +1,8 @@
-package kitchenpos.ui;
+package kitchenpos.order.ui;
 
-import kitchenpos.application.OrderService;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.application.OrderService;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.request.OrderCreateRequest;
 import kitchenpos.response.OrderResponse;
 import org.springframework.http.ResponseEntity;
