+++ src/main/java/kitchenpos/response/OrderLineItemResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.response;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 import java.util.List;
 import java.util.stream.Collectors;
